Tweak TextRegexSanitizer to better represent the behavior of isSubFormat
This change breaks backward compatibility as it inverts the default bahavior (false instead of true) - therefore, the tests were updated to reflect this.

Added simple test to see if when already sub-formatted, the logic would still work. I cannot verify if the regex is well updated to reflect the sub-formatting, but now, if the regex already contains  sub-formats, then new subformats won't be applied.

This tweak also contains a fix for when inputting the ^ and $ characters in the regex, those are now placed better in the resulting regex.